1.48. Menerapkan Authorization di Notes API - Mengubah Pemanggilan Fungsi NotesService di NotesHandler

+++ Project/notes-app-back-end/src/api/notes/handler.js
@@ -20,10 +20,15 @@ class NotesHandler {
         tags,
         body,
       } = request.payload;
+      const {
+        id: credentialId,
+      } = request.auth.credentials;
+
       const noteId = await this._service.addNote({
         title,
         tags,
         body,
+        owner: credentialId,
       });
 
       const response = h.response({
@@ -56,8 +61,12 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const {
+      id: credentialId,
+    } = request.auth.credentials;
+
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -71,6 +80,11 @@ class NotesHandler {
       const {
         id,
       } = request.params;
+      const {
+        id: credentialId,
+      } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
 
       return {
@@ -106,7 +120,11 @@ class NotesHandler {
       const {
         id,
       } = request.params;
+      const {
+        id: credentialId,
+      } = request.auth.credentials;
 
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.editNoteById(id, request.payload);
 
       return {
@@ -139,6 +157,11 @@ class NotesHandler {
       const {
         id,
       } = request.params;
+      const {
+        id: credentialId,
+      } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.deleteNoteById(id);
 
       return {
